pyproject.toml
[tool.poetry]
name = "isagog-userauth"
version = "0.1.0"
description = "A FastAPI project to manage users and issue JWT tokens to control access"
authors = ["Robert Alexander <gogonegro@gmail.com>"]
readme = "README.md"
#packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
